+++ Tests/ChessUIHarness/ChessUIHarness/ChessUIHarnessApp.swift
@@ -1,0 +1,17 @@
+//
+//  ChessUIHarnessApp.swift
+//  ChessUIHarness
+//
+//  Copyright © 2026 Päike Mikrosüsteemid OÜ. All rights reserved.
+//
+
+import SwiftUI
+
+@main
+struct ChessUIHarnessApp: App {
+    var body: some Scene {
+        WindowGroup {
+            HarnessView()
+        }
+    }
+}